First phrase, 'a' protocol
This resolves my informal objection (https://lists.w3.org/Archives/Public/public-new-work/2023Oct/0003.html)
in which I argued the name of the WG
should be a question and not an answer.

Let's keep the short name "Solid WG" as it is, but read it as a question instead
of reading it as an answer. The first phrase of the charter then
explains the title. In this case, by saying "a protocol" instead of "the protocol",
it explains that "Solid WG" means
"the WG about Social-Linked-Data as a question that need answering" and not
"the WG about the specific pre-selected answer that we developed so far".