change name and mission of the WG
This is a big PR (sorry).

I changed the name of the WG and deliverable to PUMPKIN (working title).

I mostly touched the first sections (Mission, Motivation & Background,
Scope, Deliverables).
I refactored the text to make it more about the problem and less about
the proposed solution. For the most, it is a reorganization of text that
was already there, sometimes slightly rephrased, sometimes moved around
(especially between the Motivation and Scope section).

The text that is really new compared to the previous version is
highlighted in green to make it easier to review.